requirements.txt
streamlit==1.50.0
pandas==2.3.3
plotly==5.24.1
openpyxl==3.1.5
requests==2.32.5
Pillow==11.3.0
scrapy==2.12.0
playwright==1.50.0
python-docx==1.2.0
lxml==6.0.2
markdown==3.9
